+++ DTOs/CreateVacancyDTO.cs
@@ -4,7 +4,7 @@ namespace FindJobWebApi.DTOs
 {
     public class CreateVacancyDTO
     {
-        public int CompanyId { get; set; }
+        public ulong CompanyId { get; set; }
         [Required]
         public string Title { get; set; }
         [Required]
